pyproject.toml
[project]
name = "personal-ai-stock-manager"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "pyyaml>=6.0",
  "pydantic>=2.0",
  "pytest>=8.0",
  "requests>=2.31",
  "pandas>=2.0",
  "numpy>=1.26",
  "mcp>=1.10.0",
  "boto3>=1.34",
  "jsonschema>=4.20",
  "streamlit>=1.36",
  "altair>=5.0",
  "plotly>=5.20",
  "deepdiff>=7.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
